% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedFamilias-package.R
\docType{package}
\name{pedFamilias-package}
\alias{pedFamilias}
\alias{pedFamilias-package}
\title{pedFamilias: Convert Pedigree Formats Between Pedsuite and Familias}
\description{
Read and write .fam files associated with the 'Familias' software, based on the pedigree format used in the 'pedsuite'.
}
\author{
\strong{Maintainer}: Magnus Dehli Vigeland \email{m.d.vigeland@medisin.uio.no} (\href{https://orcid.org/0000-0002-9134-4962}{ORCID})

}
\keyword{internal}
