% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFam.R
\name{writeFam}
\alias{writeFam}
\alias{openFamilias}
\title{Export data to .fam}
\usage{
writeFam(
  ...,
  famfile = "ped.fam",
  params = NULL,
  dbOnly = FALSE,
  openFam = FALSE,
  FamiliasPath = NULL,
  verbose = TRUE
)

openFamilias(famfile = NULL, FamiliasPath = NULL, verbose = TRUE)
}
\arguments{
\item{...}{For ordinary files, one or several pedigrees. Each argument should be either
a single \code{ped} object or a list of such. If the pedigrees are unnamed, they are
assigned names "Ped 1", "Ped 2", etc. For DVI files, a single list with entries \code{pm},
\code{am} and \code{missing}.}

\item{famfile}{The name or path to the output file to be written. The extension ".fam"
is added if missing.}

\item{params}{A list of Familias parameters controlling database settings, dropout and
other options. See Details.}

\item{dbOnly}{A logical. If TRUE, no pedigree information is included; only the
frequency database.}

\item{openFam}{A logical. If TRUE, an attempt is made to open the produced \code{.fam} file
in an external Familias session. Only available on Windows systems with a working
Familias installation.}

\item{FamiliasPath}{The path to the Familias executable. If empty, the following are
tried in order: "Familias3.exe", "C:/Program Files (x86)/Familias3/Familias3.exe".}

\item{verbose}{A logical, by default TRUE.}
}
\value{
The file name is returned invisibly.
}
\description{
This function produces a \code{.fam} file readable by the Familias software (Egeland et al.,
2000), containing all input pedigrees, their marker data and mutation models. The
option \code{openFam = TRUE} calls \code{openFamilias()} to open a fresh Familias session with
the produced file loaded.
}
\details{
\strong{Parameters}

The \code{params} argument controls various Familias settings. Most entries below  may be
omitted, in which case suitable defaults are used. Single values are recycled over
markers or individuals as appropriate. The \code{params} argument is broadly compatible with
the \code{params} component produced by \code{\link[=readFam]{readFam()}} with \code{includeParams = TRUE}.

\emph{General parameters}
\itemize{
\item \code{version}: Familias version written to the file. Default: \code{"3.4.1"}.
\item \code{dvi}: Logical indicating whether to write a file for the Familias DVI module.
Default: \code{FALSE}.
}

\emph{Database parameters}
\itemize{
\item \code{dbName}: Name of the marker database. Default: \code{"unknown"}.
\item \code{dbSize}: Database size for each marker. Default: \code{1000}.
\item \code{maf}: Minor allele frequency for each marker. Default: \code{0}.
\item \code{theta}: Theta/kinship/Fst correction. Default: \code{0}.
\item \code{dropoutValue}: Dropout probability for each marker. Default: \code{0}.
\item \code{dropoutConsider}: Logical vector indicating which pedigree members should be
subject to dropout.
\item \code{dropout}: Convenience alternative to \code{dropoutConsider} and \code{dropoutValue}; either a
single dropout probability or a named vector identifying individuals with dropout. For
DVI files \code{dropout} has a different meaning (see below).
}

\emph{DVI parameters}

The following probabilities pertain only to Direct/Identity matching in the DVI module:
\itemize{
\item \code{dropout}: Dropout probability. Default: \code{0}.
\item \code{dropin}: Dropin probability. Default: \code{0}.
\item \code{typingError}: Typing error probability. Default: \code{0}.
}

\strong{DVI data}

For DVI files, set \code{params$dvi = TRUE} and supply a single list with the following
entries:
\itemize{
\item \code{pm}: A list of singletons; the victim samples.
\item \code{am}: A named list of reference families. Each must be a connected \code{ped} object, and have
exactly one missing person.
\item \code{missing}: A vector naming the missing persons.
}

To ensure that the input data has the correct format, we recommend writing DVI files
through the wrapper \code{dvir::writeFamDVI()}.
}
\examples{

# Create pedigree with 2 markers
x = nuclearPed() |>
  addMarker(geno = c("2/2", "1/3", "2/3"), alleles = 1:3,
            afreq = c(.3,.3,.4), name = "M1")

# Write to .fam
tmp = writeFam(x, famfile = tempfile())

# Read back in
y = readFam(tmp)

stopifnot(identical(x, y))


### With stepwise mutation model
x2 = setMutmod(x, model = "stepwise",
               rate = list(male = 0.001, female = 0.002),
               range = 0.1, rate2 = 0.0001)

# Write and read
y2 = x2 |>
  writeFam(famfile = tempfile()) |>
  readFam()

stopifnot(identical(x2, y2))


### Read/write including detailed parameters
params = list(theta = 0.1, dbName = "myDB", dropout = c("3" = 0.01))
fam = writeFam(x2, famfile = tempfile(), params = params)

dat = readFam(fam, includeParams = TRUE)

# Pedigree is now in the `main` slot
stopifnot(identical(x2, dat$main))

# The `dropout` parameter is converted to (and is equivalent to):
dat$params$dropoutConsider
dat$params$dropoutValue


### Read/write frequency database

# Write database as fam file
dbfam = writeFam(x2, famfile = tempfile(), dbOnly = TRUE)

# Read back in: The result is a list of marker attributes
a = readFam(dbfam)

# Attach to a pedigree and write to a new file
z = singleton(1) |> setMarkers(locusAttributes = a)
dbfam2 = writeFam(z, famfile = tempfile(), dbOnly = TRUE)

stopifnot(identical(readLines(dbfam), readLines(dbfam2)))


### DVI file

pm = singleton("V1") |> addMarker(geno = "1/2", name = "M")

am = nuclearPed(children = "MP") |>
  addMarker(geno = c("1/1", "1/2", NA), name = "M")

x = list(pm = list(V1 = pm), am = list(F1 = am), missing = "MP")
writeFam(x, famfile = tempfile(fileext = ".fam"), params = list(dvi = TRUE))

}
\references{
Egeland et al. (2000). \emph{Beyond traditional paternity and identification
cases. Selecting the most probable pedigree.} Forensic Sci Int 110(1): 47-59.
}
\seealso{
\code{\link[=readFam]{readFam()}}.
}
