% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedFamilias-package.R
\docType{package}
\name{pedFamilias-package}
\alias{pedFamilias}
\alias{pedFamilias-package}
\title{pedFamilias: Import and Export 'Familias' Files}
\description{
Tools for exchanging pedigree data between the 'pedsuite' packages and the 'Familias' software for forensic kinship computations (Egeland et al. (2000) \doi{10.1016/s0379-0738(00)00147-x}). These functions were split out from the 'forrel' package to streamline maintenance and provide a lightweight alternative for packages otherwise independent of 'forrel'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/magnusdv/pedFamilias}
  \item Report bugs at \url{https://github.com/magnusdv/pedFamilias/issues}
}

}
\author{
\strong{Maintainer}: Magnus Dehli Vigeland \email{m.d.vigeland@medisin.uio.no} (\href{https://orcid.org/0000-0002-9134-4962}{ORCID})

}
\keyword{internal}
